Add a switch for hacky `current_results` table gen

+++ sample_work/sample_work.py
@@ -38,15 +38,16 @@ def organisation_ids():
     return query('select id from organisation;')
 
 class WorkItems(object):
-    def __init__(self, org_ids, test_ids):
+    def __init__(self, org_ids, test_ids, regen_results):
         self.org_ids = org_ids
         self.test_ids = test_ids
-        query('''DROP TABLE IF EXISTS current_result''', write=True)
-        query('''CREATE TABLE current_result AS
-                   SELECT * FROM result
-                   WHERE test_id = ANY(%s)
-                   AND organisation_id = ANY(%s)
-                   AND result_data > 0''', (test_ids, org_ids,), write=True)
+        if regen_results:
+            query('''DROP TABLE IF EXISTS current_result''', write=True)
+            query('''CREATE TABLE current_result AS
+                       SELECT * FROM result
+                       WHERE test_id = ANY(%s)
+                       AND organisation_id = ANY(%s)
+                       AND result_data > 0''', (test_ids, org_ids,), write=True)
 
     def test_string_of_test_id(self, test_id):
         results = query('''select name from test where id = %s;''', (test_id,));
